+++ conftest.py
@@ -1,0 +1,4 @@
+import sys
+
+if sys.version_info.major < 3:
+    collect_ignore = ["tests/test_async_api.py"]
