+++ tox.ini
@@ -15,6 +15,7 @@ distshare = dist
 [testenv]
 deps =
     -r test_requirements.txt
+    py37: .[async]
 commands =
     pytest tests
 
@@ -22,7 +23,7 @@ commands =
 basepython = python3
 skip_install = true
 deps = pre-commit
-commands = pre-commit run --all-files --show-diff-on-failure
+commands = pre-commit run --all-files --show-diff-on-failure  # todo add coverage
 
 [testenv:build]
 skip_install = true
